dev-manual, sdk-manual: Fixed link to Eclipse section
The YP doc set was using a link to an out-dated video that
showed how to configure and run Eclipse.  The video was very
old and Jessica suggested just removing it.  So, I replaced
all occurrences of the link to the up-to-date appendix in the
SDK manual that provides information on the latest supported
version of the Eclipse IDE.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>