yocto-project-qs: Updated the YP Development Flow figure
The figure that Scott Garman used for an old video cast is
much more appealing to the eye than the giant, square beast
the current manual was using.  I have replaced the image.
The image is technically the same.

Because the mega-manual has to use a duplicate figure, I
put the new PNG files in the Figures directory for both the
yocto-project-qs and mega-manual books.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>