dev-manual, ref-manual: Reorganized Bugzilla information
Fixes [YOCTO #11630]

Reorganized the way the information about entering bugs using
Bugzilla is arranged in the documentation set.  The dev-manual now
has a section that is purely procedural and steps the user through
the process of logging a new bug in the Bugzilla system.  The
ref-manual has a conceptual section that introduces the YP
implementation of Bugzilla by simply telling the user what it is
and what what its purpose is.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>