dev-manual, ref-manual: Reorganized "Other Information"
Fixes [YOCTO #11630]

The exhaustive list of useful links and references to other
YP documentation existed in the dev-manual.  This is not the correct
place for this type of reference information.  I have moved that
list to the ref-manual where it should be.  The dev-manual now
simply provides a link into that section.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>