documentation: Fixed links to "bitbake-term"
Fixes [YOCTO #11630]

Moving the "Yocto Project Terms" section from the dev-manual to
the ref-manual.  Doing so caused all the links to the id
"bitbake-term" to break.  These had to be individually fixed.

Discovered two unresolved references that were a consequence of
moving that section to the ref-manual.  These were fixed as well.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>