documentation: Reorganized YP introduction material.
This material was being replicated across the Quick Start,
dev-manual, and ref-manual.  I have created a single point in the
ref-manual that collects the disparate information that existed.

I deleted the entire section in the dev-manual since that book
is converting to a procedure manual and the very front of the manual
adequately mentions YP and provides pointers to where the user can
get more information.

The QS provides a more complete introduction of YP since it is a
QS.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>