dev-manual, ref-manual: Added section on index of releases procedure
Fixes [YOCTO #11630]

I have added a new procedure on using the Index of Releases area
to get snapshots of various tarballs.  I recommend that is not
the way you usually want to get files as using Git is recommended.
The section exists for completeness.

In the ref-manual, I cross-reference the procedures from the
associated conceptual areas.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>