kernel-dev, dev-manual: Moved Yocto Linux kernel overview
Fixes [YOCTO #11630]

I took the Yocto Linux kernel overview list of kernels out of the
dev-manual and moved them to the "Yocto Project Kernel Development
and Maintenance" section in Appendix A of the kernel-dev manual.
This is a much more logical area to introduce this type of Yocto
Linux kernel information.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>