dev-manual, kernel-dev: Moved kernel branch concepts to kernel-dev
Fixes [YOCTO #11630]

The information in the dev-manual kernel overview area really neeeds
to be in the Appendix on kernel structure in the kernel-dev manual.
I moved that informtaion to the appendix.  Removal of one redundant
image was necessary from the dev-manual.  The figure was literally
repeated in the kernel manual already under a different file name.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>