kernel-dev, dev-manual: Moved Kernel workflow section to kernel-dev
Fixes [YOCTO #11630]

The overview section of modifying the kernel needed to be in the
kernel-dev manual and not in the dev-manual.  I moved it and took
care of all the links and external references throughout the YP
doc set.  One figure was involved that had to be moved to the figures
folder in kernel-dev from dev-manual.  That is why the Makefile had
to be edited.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>