dev-manual: Updated "Performing a Simple Build" section.
I added a figure for this simple build workflow that uses
BitBake from a Linux host.  Also did a bit of re-writing.

A new figure was introduced so the Makefile had to be adjusted
to include that figure.  The figure resides in the figures
directory of both the dev-manual and the mega-manual.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>