dev-manual, ref-manual: Eliminated pre-built section
Fixes [YOCTO #11630]

I took the section that described how to use a pre-built kernel
and run it through QEMU out.  This is basically a QEMU usage
section and is not in that area.  There were some QEMU speed
up items suitable for the QEMU concepts section in the ref-manual.
I put those in that area.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>