yocto-project-qs, ref-manual: Moved YP intro from QS to ref-manual
Fixes [YOCTO #11630]

The introductory stuff in the QS is really conceptual information and
needs to be in the ref-manual where YP is introduced.  Regarding the
QS, all we really need to do is point to places where the reader can
go to find that stuff out.

Part of this move involved getting the diagram of the YP flow for
the environment from the QS to the ref-manual.  That figure was
named "YP-flow-diagram.png".  It was named "yocto-environment.png"
in the QS but I renamed it when moving it to the ref-manual.  This
caused some edits to the "Makefile" to clean up the figure lists
for tarballs.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>