kernel-dev, dev-manual: Moved Kernel prep steps
The dev-manual had a place-holder section for getting the
build host ready for using YP and then doing kernel
development.  This stuff is better suited for the actual
kernel development manual.  I moved the section.  Moving
created several broken links in the manual set that had to
be fixed.

In the dev-manual, I updated the introductory list to include
the newly moved kernel stuff.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>