ref-manual: Fixed YP Term problem with botched earlier commit
I was cherry-picking in commits from master to pyro and had a
conflict that I did not go far enough to the bottom of the
file to see the true nature, which was duplication of the
"Yocto Project Terms" section.  When I resolved the conflit
I just took out the top couple lines and actually left the
duplicated terms section in.  Then I pushed everthing.  I should
have made the manuals first and I would have discovered the
error.

This commit fixes it.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>