sdk-manual: Updated Eclipse information
Fixed [YOCTO #11970]

Updated the field description for "Specify Sysroot Location" to
note that the user can have this either downloaded or built and
the ramifications of using a downloaded root filesystem.

Updated the "Deploying and Debugging the Application" steps for
both supported Eclipse versions to tell the user what to do if
a browse operation fails or times out.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>