ref-manual: Added "N" category to glossary
Discovered this item in the front of the glossary was commented
out and not showing up so the user did not have the option to
click on "N" to quickly get to the "N" entries, which did exist
and show up.  Uncommented it and put the proper beginning "N"
variable in as the start.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>