dev-manual, ref-manual: Image format implementation changes
Recently a number of changes have happened on the implementation
side of some image formats, and a few have been renamed.
The u-boot image signing code is now always available and no
longer in a stand-alone file.  The vmdk/mdi/qcow2 images have
been removed and are now just a conversion type that is applied
to wic images.

Reported-by: Tom Rini <trini@konsulko.com>
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>