kernel-dev: Eliminated "Generating Configuration Files" section.
This section was largely redundant now.  The small things that did
add value I placed in other areas of the chapter 2.  One was a new
section on expanding variables using bitbake -e.  Another was a bit
about the defconfig file.  And a small procedure on how to create
config fragments using the "diffconfig" command was retained.

Links from the dev-manual and the ref-manual were adjusted as well.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>