dev-manual, ref-manual: Moved kickstart reference to ref-manual
This section was misplaced in the dev-manual and really needed to
be in the ref-manual.  I created a new chapter that follows the
devtool reference chapter and put the kickstart chapter there.

Links were needed to be fixed up and a new chapter in the
ref-manual named ref-kickstart.xml had to be created.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>