yocto-project-qs: Updated some links to work better from mega-manual
Links to the top of a manual don't really work from the mega-manual.
I updated the manual links at the bottom of the yocto-project-qs
so that they go to the introduction sections of the appropriate
manual rather than attempting to go to the top of the manual, where
they don't render from within the mega-manual.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>