documentation: Updates to cull out buried YP Website references.
With the redoing of the YP website, it was decided that the manuals
could only reference into the top-level and the top-level/documentation
URLs of the website.  I removed all references to areas such as
projects (e.g. Toaster, VM, etc.) and any other URL that was not
the top-level landing page or the documentation page.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>