sdk-manual: Split the Eclipse section into a separate chapter
This also affected a link in the dev-manual and some setup stuff
in the mega-manual and sdk-manual.xml file due to a new chapter
being introduced.

This section on the Eclipse stuff is big enough to warrant its
own chapter.  I decided to pull it into a separate chapter of its
own.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>