dev-manual, kernel-dev: Replaced yocto-layers with bitbake-layers command
The yocto-layers script is being deprecated in the 2.4 YP release
and the preferred script is now bitbake-layers.  In particular, the
section that talks about creating a general layer was using
yocto-layers create.  I have replaced it with bitbake-layers
create-layer command.

The bsp-guide and kernel-dev manual had links that used (in the
case of kernel-dev) and referenced yocto-layers.  I had to change
these to reflect using bitbake-layers create-layers.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>