dev-manual, ref-manual: Moved source file sections
The sections in the introductory chapter of the ref-manual are better
served as part of the dev-manual.  These sections are about accessing
YP source files using various methods (i.e. source repositories,
index of releases, and nightly builds).  I have integrated that
information into the dev-manual where the information was also being
treated.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>