overview-manual: Added new "overview" manual to YP doc set
Fixes [YOCTO #12370]

I added all the mechanics for this new "Yocto Project Overview Manual".
For now, the manual builds and publishes into the 2.5 area.  I used
"overview-manual" for the folder name.  Makefile properly builds the
new manual.  The mega-manual.sed file processes links for the new manual.
The mega-manual.xml file includes the new manual.  The poky.ent file
includes a new ENTITY link for the overview manual.  All pieces are in
place.

With this commit, just an overview chapter exists.  The remaining content
needs to be developed.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>