mega-manual.sed: Corrected bad SDK title string
I neglected to change the string for matching purposes in the SDK
line when the SDK manual title changed.  This was causing links to
that manual where there is no specific #<target> to not be resolved
correctly.

I also removed a redundant line for the yocto-project-qs manual.  It
was appearing in two places.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>