documentation: Created new YP Overview Manual
Fixes [YOCTO #12370]

Added a new chapter to the newly created Yocto Project Overview
Manual.  This chapter originated from the YP Reference Manual and
was the old chapter 3, which talked a lot about various concepts.
This information is better suited for the new overview manual.

The change involved moving the entire chapter and renaming it.
This move of the content affected many, many external references
and links into the old area.  Consequently, I had to recast all
these links from the different manual.

Changes also included fixing the mega-manual.xml file so that it
would include the new overview manual as part of it.

Many figures had to be relocated as part of the move as well.  This
meant deleting them from the ref-manual/figures folder and putting
them into the overview-manual/figures folder.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>