ref-manual: Fixed targets for sdk-manual links
I had tried using a #-type target for some links designed to go
to the top of the sdk-manual. It was not working out so I removed
the targets. The links now render benign for the mega-manual as
they should and render hot for the individual manual as they should.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>