dev-manual, ref-manual: Moved Wic Plug-In section to dev-manual
Fixes [YOCTO #12370]

I had a section on Wick Plug-Ins in the ref-manual in the
"technical details" chapter.  This section has been combined with the
section on using Wic that lives in the dev-manual.  This move creates
a single section on using Wic to create Wic-partitioned images.

The section was moved out of the ref-manual and merged into the
dev-manual Wic section.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>