ref-manual, overview-manual, Makefile: Moved toolchain concepts
Fixes [YOCTO #12370]

Moved the "Cross-Development Toolchain Generation" section from the
ref-manual to the overview-manual into the concepts chapter.  This
information is conceptual and now needs to live in the concepts
chapter of the new overview-manual.  Moving the section caused a
few links to have to be fixed in the ref-manual.  There was also a
figure in the section.  So, I had to move the figure from the ref-manual
to the overview-manual "figures" folder and update the Makefile for
the TARFILE generation.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>