yocto-project-qs: Defined Poky and added overview manual link
Fixes [YOCTO #12128]

Updated the phonetic rendition of the term "poky".  Historically,
this term is mispronounced.  I am now saying "Pock-ee" as the official
way of pronouncing this term.

I also added a link to the Yocto Project Overview Manual that was
in the Tip box.  The manual did not exist earlier.  Since it does
exist now, I have added the link.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>