ref-manual: Removed the "technical-details.xml" file
Fixes [YOCTO #12370]

The "technical-details.xml" file was for the old chapter four of the
ref-manual.  All content in that chapter has been moved to the new
concepts chapter of the overview-manual.  I updated the mega-manual.xml
file to not include the old technical-details chapter as part of the
ref-manual.  I also updated the ref-manual.xml file to not include the
technical-details.xml chapter when building the ref-manual.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>