ref-manual, dev-manual: Moved QEMU info from ref-manual to dev-manual
Fixes [YOCTO #12370]

The QEMU information in the ref-manual "using" chapter has been moved
to the dev-manual's chapter 5, which is a dedicated QEMU chapter.
Some information is reference and that is why I had it separated out
into the reference manual.  However, pressures are mounting to get
that stuff out of the ref-manual and into the dev-manual.  So, it
has been moved.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>