documentation: Removed "usingpoky" chapter from ref-manual
Fixes [YOCTO #12370]

All of the information from the "usingpoky" chapter in the
ref-manual has been distributed out over the rest of the
YP manual set.  Primarily, this information went into the
dev-manual and the overview-manual.  Because the chapter is
no more, I had to update the mega-manual.xml to not include that
chapter. Also, had to update ref-manual to exclude the chapter
as part of the Make process.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>