getting-started, mega-manual: Added reference distro poky section
Added the new section "Reference Embedded Distribution (Poky)"
section to the getting-started manual.  The new section required
a new figure.  I had to add the figure to the figures folder of
both the getting-started and mega-manual books.  Updates to the
Makefile to include the new figure as part the tarfiles created
for each book.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>