brief-yoctoprojectqs, Makefile: New "First Build" article.
I have created an brief article in the form of current YP docs
that is a pared down version of the Quick Start (pre 2.5 release).
The article fills a gap for a short (2-page), process-oriented
article that steps the user through an initial build using YP.

I had to make changes to the Makefile to support building the new
article.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>