bsp-guide: Removed deprecated tool sections
I took out the sections at the end of the manual that talked
about the yocto-kernel tool.  This tool is no longer maintained
and there is no equivalent tool.  I also fixed the yocto-bsp
tool to be the bitbake-layers tool.  This involved some
consolidation of sections.

I fixed some links in the kernel-dev and toaster-manual.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>