ref-manual: Updated PACKAGES_FEED_ARCHS variable
Added the fact that this variable is really optional.  If  you
don't use it then you are setting up a package feed URI that
will include all supported arcchitectures for the package.
Using it means you are setting up specific URIs for specific
architectures.

Added a "Tip" note box to show how you can use the variable
to whitelist architectures.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>