dev-manual: Cleaned up layer naming terminology
Reality dictates that the YP does not treat layer naming consistently.
Layers are talked about as the "meta-whatever" layer and never the
"whatever" layer.  However, the tooling sometimes appends or uses
just the "whatever" part of the layer name.  A good example is the
meta-yocto-bsp configuration file.  All the variables in there use
just the root name of a layer's name.  In the manuals, I had been
distinguishing the layer name as just the "whatever" part sans
"meta-".  I talked about the convention of using "meta-" in front
of layer names, etc.  Well, this is confusing in light of how everyone
says a layer's name is "meta-whatever".  So, I fixed all this up.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>