bsp-guide: Updated BSP terminolgy and BBLAYERS ordering
Fixed the way we refer to a BSP name.  It is really
"meta-<bsp_root_name>" rather than "meta-<bsp_name>".  The
name is the whole string and not just the root name.

Also added a tip on ordering the layers in the BBLAYERS
variable in the bblayers.conf file.  Order is important.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>