poky.ent: Added two Minus-1 variables for release code names
Many examples in the YP set use the latest release code name
as part of a list and then the "minus-1" code name is literally
typed in.  I think it is time to have a couple ENTITY variables
in poky.ent that will fill in the last release for these examples.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>