documentation: Updated manual notes
I updated the notes to help the user get the version of the
docs that they are interested in.  Sometimes a search using the
web returns really old versions of the manual and the user
is clueless about using a manual that is not matching the
YP release they are working with.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>