documentation: Fixed "Getting Started" manual links
There was inconsistent use when referencing the getting-started
manual.  The manual's name does not include the word
"Manual".  So, I cleaned this up across the manual set and
also in the mega-manual.sed file, which was incorrectly
processing the links to that manual that do not target a
specific section.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>