sdk-manual: Removed non-eclipse option for Eclipse flow
The step in the general Eclipse IDE workflow has a step that tells
the user to "Create and Build Your Application."  I had a sentence
in that step that suggested if you were not using Eclipse to follow
some sort of procedure.  The whole idea of this flow is to be
working within Eclipse.  I removed that sentence.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>