mega-manual.sed: Added overview manual and extra subst.
Updated the script to process the overview-manual strings
found in the mega-manual.  These changed due to the folder
chaning from "getting-started" to "overview-manual".

Additionally, I discovered a single occurrence of a linked
reference to the mega-manual that was in the mega-manual.
This is not ideal.  That link needs to be a link in the
ref-manual where it originates.  However, the link in the
mega-manual should not link to itself.  That link needed
processing just like all top-level book links in the
mega-manual.  The link should be removed.  I added a single
substitution operation at the end of the file to handle this
lone case within the mega-manual.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>