ref-manual: Fixed a link to a naked manual reference
The terms.xml chapter had a link to the toaster-manual where the
string to the manual was not matching the string used in the
mega-manual.sed file.  Consequently, the link was not being processed
correctly.  This type of link should not be a link in the mega-manual
because the target is to the manual in general.  Fixed the string
of the toaster-manual to match what is being used in the mega-manual.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>