ref-manual: Provided more relevant link for Python documentation
In the BBMASK variable, a link to Python documentation was
going to a specific 2.6 release.  That release was probably the
most recent when the BBMASK variable was initially documented.
Python has many releases now (e.g. 3.6.5).  I re-worded the
description to be more generic and sent the user to the general
Python release area.  They can pick a release from that URL.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>