poky.ent: Deleted the ENTITY for the YP Release Notes
This ENTITY was used in two places.  I have updated those areas to
give the user new instructions to reach the Release Notes.  As of
YP 2.5, we do not have a simple URL that can be used for the
Release Notes.  The user must go to the "DOWNLOADS" page of the
main YP website and click the "RELEASE INFORMATION" button.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>