ref-manual: Updated the sections pointing to the YP Release Notes.
The method for accessing the YP Release Notes for YP 2.5 forward
is to go to the "DOWNLOADS" page of the main YP website and then
click on the "RELEASE INFORMATION" button.  No longer does a user
need a simple URL.  I updated the two spots in manual where this
information existed for accessing Release Notes.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>