dev-manual: Updated managing granular level packaging discussion.
In the section that talks about building an image for multiple
machines, there was a bullet item on managing granular level
packaging.  It had an old example that used the "daisy" release
that showed how to share packages and re-use binaries.  I replaced
the example with a discussion of Freescale's "fsl-dynamic-packagearch"
class, which does this type of sharing.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>