dev-manual: New section in QEMU for workaround under KVM
As suggested in a recent mail thread by Martin Jansa, there are cases in
which QEMU under KVM will crash because of CPU feature incompatibilities
between compiled binaries and the host CPU under which qemu is run.
Although this is hard to fully escape, we should document the issue to
help people work around it.  I have taked Martin's suggested patch
and did some rewording for the new section.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>