ref-manual: Updated PACKAGE_ARCH variable description.
I added a link to the BUILD_ARCH variable.  The variable is part
of the glossary but for some reason there was no link to it.
I also added a note for a link to the SDK_ARCH variable, which
was described inline.  By convention, I do not like to link to
variables inline as the underlining used to indicate a link messes
with with underscore characters that are usually a part of a
varible's name.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>