overview-manual: Updated "Patching" section.
This section presents a deep dive into the patching phase of the
build process.  I removed the vague "apply=yes" reference and added
more references for the reader to read and learn about patching.

The bit about the SRC_URI parameters (i.e. "apply") is going to be
covered in the do_patch documentation since that is the task that
handles that stuff.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>