dev-manual: Moved the "Simple Build" section
I had a section on building an image up in chapter that talks about
getting set up to use YP.  I moved the build section to a new parent
section that will hold topics on building various things.  I renamed
the section to "Building a Simple Image".  I had to fix several links
in the sdk-manual, overview-manual, and dev-manual.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>