dev-manual: Removed "newbie" chapter.
This chapter only had the stuff about submitting a defect against
YP and making a change to YP.  I moved that information into the
common tasks chapter as it is "how-to" information.  Removal of this
chapter alse required that the mega-manual.xml file be updated to
not include the chapter.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>