sdk-manual: Changed "configure.ac" file in the autotools-based example
The line used to be "AC_PROG_INSTALL".  I did some digging and can't
really figure out why the example is not following the standard
"helloworld" stuff out there all over the internet.  So, I have the
user create a configure.in file as follows now:

   AC_INIT(hello,0.1)
   AM_INIT_AUTOMAKE([foreign])
   AC_PROG_CC
   AC_CONFIG_FILES(makefile)
   AC_OUTPUT

The original was as follows:

   AC_INIT(HELLO,0.1)
   AM_INIT_AUTOMAKE([foreign])
   AC_PROG_CC
   AC_PROG_INSTALL
   AC_OUTPUT(Makefile)

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>