sdk-manual: Added cross-reference to Makefile section. Fixed syntax
Two changes here.  One was a note to the "make" step of the
Autotools-based section.  I cross-referenced the Makefile section
for information on how SDK installation environment variables
are respected and or overridden when using make variables.

Also, fixed the quotation syntax of the four environment variable
examples used in the "Makefile-Based Projects" section.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>