sdk-manual: Updates to Eclipse Workflow section
Updated the "Workflow Using Eclipse" section head to not match so
closely the "Working With Eclipse" section heading.  TOC appeared
funny.

Also, updated the step 1 to be more accurate regarding the real
need of having a YP build host and where to go to get information
on setting it up.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>