sdk-manual: Updates to sections for updating installed Ext SDKs
There are two sections that deal with this topic.  One is from the
standpoint of a user using a 3rd party published and installed
extensible SDK ("Applying Updates to an Installed Extensible SDK").
The other is from the standpoint of a person providing Extensible
SDKs for consumption ("Providing Updates to the Extensible SDK
After Installation").  These sections needed some closer examination
regarding fully describing what was going on.  I provided some
re-writes to both.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>