ref-manual: Added KERNEL_ARTIFACT_NAME and adjusted referencing variables.
The KERNEL_ARTIFACT_NAME variable is used throughout to set the names
of build artifacts.  Rather than repeat informaiton about
KERNEL_ARTIFACT_NAME in the many variables that use it, I added
a new entry for the variable.  This also impacted the descriptions
of the variables that were repeating information.  I updated those
variable descriptions as well.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>