brief-yoctoprojectqs, dev-manual: Updated poky clone examples.
The examples in these manuals for checkout of poky by tag need
to be updated immediately after a release.  The reason is that the
examples use recent tags.  I have updated both the examples
that show how to checkout poky based on a specific tag.  This
particular commit makes the YP 2.5.2 release examples correct.
Note, that I use &DISTRO; ENTITY variables to do this but timing
for when they are correct for a release needs to be monitored.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>