dev-manual, brief-yoctoprojectqs: Added "2.6.1" tag to examples
Even though the tag does not exist at this point because the
release is in development, I added the tag to the list of
tags returned by the "git tag" command in the two examples
in these books.  The manual at this point is under development
so all bets are off as to things being totally accurate here.
Once the release is out, the example will work.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>